requirements.txt
beautifulsoup4
Django
djangorestframework
requests
